backend/requirements.api.txt
fastapi>=0.104.1
slowapi>=0.1.9
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
redis>=5.0.0
SQLAlchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=1.1.3
httpx==0.27.2
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-multipart>=0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
pypdf>=3.17.0
python-docx>=1.1.0
nltk==3.8.1
scikit-learn>=1.3.0
anyio>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
openai==1.40.0
extruct>=0.16.0
w3lib>=2.1.2
readability-lxml>=0.8.1
reportlab>=4.0.0
html2docx>=1.1.0
bcrypt>=4.1.0
apscheduler>=3.10.0
gunicorn==21.2.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
python-magic==0.4.27
pyotp>=2.9.0
qrcode[pil]>=7.4.2
